rename InputCommon/ControllerInterface/Device to CoreDevice
