+++ Source/Core/DolphinQt/Config/Mapping/IOWindow.h
@@ -12,7 +12,7 @@
 #include <QSyntaxHighlighter>
 
 #include "Common/Flag.h"
-#include "InputCommon/ControllerInterface/Device.h"
+#include "InputCommon/ControllerInterface/CoreDevice.h"
 
 class ControlReference;
 class MappingWidget;
