+++ Source/Core/DolphinQt/Config/Mapping/MappingWindow.cpp
@@ -47,7 +47,7 @@
 
 #include "InputCommon/ControllerEmu/ControllerEmu.h"
 #include "InputCommon/ControllerInterface/ControllerInterface.h"
-#include "InputCommon/ControllerInterface/Device.h"
+#include "InputCommon/ControllerInterface/CoreDevice.h"
 #include "InputCommon/InputConfig.h"
 
 constexpr const char* PROFILES_DIR = "Profiles/";
