+++ Source/Core/DolphinQt/Config/Mapping/MappingWindow.h
@@ -8,7 +8,7 @@
 #include <QString>
 #include <memory>
 
-#include "InputCommon/ControllerInterface/Device.h"
+#include "InputCommon/ControllerInterface/CoreDevice.h"
 
 namespace ControllerEmu
 {
