+++ Source/Core/InputCommon/ControllerEmu/ControlGroup/Cursor.h
@@ -8,7 +8,7 @@
 #include <string>
 
 #include "InputCommon/ControllerEmu/StickGate.h"
-#include "InputCommon/ControllerInterface/Device.h"
+#include "InputCommon/ControllerInterface/CoreDevice.h"
 
 namespace ControllerEmu
 {
