+++ Source/Core/InputCommon/ControllerEmu/ControlGroup/Tilt.h
@@ -8,7 +8,7 @@
 
 #include "InputCommon/ControllerEmu/Setting/NumericSetting.h"
 #include "InputCommon/ControllerEmu/StickGate.h"
-#include "InputCommon/ControllerInterface/Device.h"
+#include "InputCommon/ControllerInterface/CoreDevice.h"
 
 namespace ControllerEmu
 {
