+++ Source/Core/InputCommon/ControllerInterface/CoreDevice.h
