+++ Source/Core/InputCommon/ControllerInterface/DInput/DInputJoystick.h
@@ -4,7 +4,7 @@
 
 #pragma once
 
-#include "InputCommon/ControllerInterface/Device.h"
+#include "InputCommon/ControllerInterface/CoreDevice.h"
 #include "InputCommon/ControllerInterface/ForceFeedback/ForceFeedbackDevice.h"
 
 namespace ciface::DInput
