+++ Source/Core/InputCommon/ControllerInterface/OSX/OSXJoystick.h
@@ -8,7 +8,7 @@
 
 #include <IOKit/hid/IOHIDLib.h>
 
-#include "InputCommon/ControllerInterface/Device.h"
+#include "InputCommon/ControllerInterface/CoreDevice.h"
 #include "InputCommon/ControllerInterface/ForceFeedback/ForceFeedbackDevice.h"
 
 namespace ciface::OSX
