+++ Source/Core/InputCommon/ControllerInterface/Wiimote/WiimoteController.h
@@ -16,7 +16,7 @@
 #include "Core/HW/WiimoteEmu/Extension/Nunchuk.h"
 #include "Core/HW/WiimoteEmu/MotionPlus.h"
 #include "Core/HW/WiimoteReal/WiimoteReal.h"
-#include "InputCommon/ControllerInterface/Device.h"
+#include "InputCommon/ControllerInterface/CoreDevice.h"
 
 namespace ciface::WiimoteController
 {
